requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
certifi==2025.1.31
    # via requests
charset-normalizer==3.4.1
    # via requests
greenlet==3.1.1
    # via playwright
idna==3.10
    # via requests
playwright==1.51.0
    # via -r requirements.in
pydantic==2.11.0
    # via -r requirements.in
pydantic-core==2.33.0
    # via pydantic
pyee==12.1.1
    # via playwright
pyyaml==6.0.2
    # via -r requirements.in
requests==2.32.3
    # via -r requirements.in
typing-extensions==4.13.0
    # via
    #   pydantic
    #   pydantic-core
    #   pyee
    #   typing-inspection
typing-inspection==0.4.0
    # via pydantic
urllib3==2.3.0
    # via requests
